Cast to int
If we get there, it is either because a new enum has been added, but
not updated here. In this case we could try to print the enum. The
other, potentially more likely cause is, that the user casted an int
that was out of range. In which case we can only print the integer
value.

+++ src/mesh/difops.cxx
@@ -547,7 +547,8 @@ Field3D Div_par_K_Grad_par_mod(const Field3D& Kin, const Field3D& fin, Field3D& 
         Kin, fin, flow_ylow, bndry_flux);
   }
   throw BoutException(
-      "Unknown method `{}` - choose from `Original`, `ProductJK` or `Harmonic`.", method);
+      "Unknown method `{}` - choose from `Original`, `ProductJK` or `Harmonic`.",
+      static_cast<int>(method));
 }
 /*******************************************************************************
 * Delp2
